# Extend trace filtering to more operation types